requirements.txt
paramiko==3.4.0
python-dotenv==1.0.1
requests==2.32.3
# LangGraph & CrewAI are optional but recommended
langgraph==0.2.33
crewai==0.51.1
# for diffs
unidiff==0.7.5
